Added api endpoint to return model labels list
I added a `/model/labels` endpoint that returns the list of labels and thier id that the model could return.

This list allows application developers access to what potential labels can be returned by a prediction instead of being restricted to the list of labels returned by a single prediction.